python: move MakeVars class into separate module
... so I can reuse it.
